backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.3.0
boto3==1.16.13
botocore==1.19.13
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
Django==3.1.2
django-environ==0.4.5
django-multiselectfield==0.1.12
django-taggit==1.3.0
factory-boy==3.1.0
Faker==4.14.2
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
packaging==20.4
pickle-mixin==1.0.2
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pybase64==1.0.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.11
wrapt==1.12.1
